requirements.txt
pygithub
openai